+++ exonum-java-binding/common/src/test/java/com/exonum/binding/common/proofs/list/ListProofStructureValidatorTest.java
@@ -101,6 +101,15 @@ class ListProofStructureValidatorTest {
   }
 
   @Test
+  void visit_ProofOfAbsence() {
+    ListProofNode root = new ListProofOfAbsence(H1);
+
+    validator = createListProofStructureValidator(root);
+
+    assertTrue(validator.isValid());
+  }
+
+  @Test
   void visit_InvalidTreeHasNoElements() {
     ListProofNode left = new ListProofHashNode(H1);
     ListProofBranch root = new ListProofBranch(left, null);
@@ -216,6 +225,22 @@ class ListProofStructureValidatorTest {
     validator = createListProofStructureValidator(root);
 
     assertFalse(validator.isValid());
+  }
+
+  @Test
+  void visit_InvalidProofOfAbsence() {
+    ListProofBranch root = new ListProofBranch(
+        new ListProofBranch(
+            leafOf(V1),
+            // Having proof of absence not as a proof tree root is not allowed
+            new ListProofOfAbsence(H1)
+        ),
+        new ListProofHashNode(H2)
+    );
+
+    validator = createListProofStructureValidator(root);
+
+    assertThat(validator.getProofStatus(), is(ListProofStatus.INVALID_PROOF_OF_ABSENCE));
   }
 
   private ListProofStructureValidator createListProofStructureValidator(ListProofNode listProof) {