+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/storage/indices/ProofListIndexProxy.java
@@ -16,9 +16,8 @@
 
 package com.exonum.binding.core.storage.indices;
 
-import static com.exonum.binding.core.storage.indices.StoragePreconditions.checkElementIndex;
 import static com.exonum.binding.core.storage.indices.StoragePreconditions.checkIndexType;
-import static com.exonum.binding.core.storage.indices.StoragePreconditions.checkPositionIndex;
+import static com.exonum.binding.core.storage.indices.StoragePreconditions.checkRange;
 
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.proofs.list.ListProof;
@@ -173,15 +172,14 @@ public final class ProofListIndexProxy<E> extends AbstractListIndexProxy<E>
   }
 
   /**
-   * Returns a proof that an element exists at the specified index in this list.
+   * Returns a proof of either existence or absence of an element at the specified index
+   * in this list.
    *
    * @param index the element index
    * @throws IndexOutOfBoundsException if the index is invalid
    * @throws IllegalStateException if this list is not valid
    */
   public UncheckedListProof getProof(long index) {
-    checkElementIndex(index, size());
-
     ListProof listProof = nativeGetProof(getNativeHandle(), index);
     return new UncheckedListProofAdapter<>(listProof, this.serializer);
   }
@@ -189,7 +187,9 @@ public final class ProofListIndexProxy<E> extends AbstractListIndexProxy<E>
   private native ListProof nativeGetProof(long nativeHandle, long index);
 
   /**
-   * Returns a proof that some elements exist in the specified range in this list.
+   * Returns a proof of either existence or absence of some elements in the specified range
+   * in this list. If some elements are present in the list, but some — are not (i.e., the
+   * requested range exceeds its size), a proof of absence is returned.
    *
    * @param from the index of the first element
    * @param to the index after the last element
@@ -197,10 +197,8 @@ public final class ProofListIndexProxy<E> extends AbstractListIndexProxy<E>
    * @throws IllegalStateException if this list is not valid
    */
   public UncheckedListProof getRangeProof(long from, long to) {
-    long size = size();
-    ListProof listProof = nativeGetRangeProof(getNativeHandle(),
-        checkElementIndex(from, size),
-        checkPositionIndex(to, size));
+    checkRange(from, to);
+    ListProof listProof = nativeGetRangeProof(getNativeHandle(), from, to);
 
     return new UncheckedListProofAdapter<>(listProof, this.serializer);
   }